feat: jcr invalidation implementation
