+++ pom.xml
@@ -22,7 +22,7 @@
     <parent>
         <artifactId>jahia-modules</artifactId>
         <groupId>org.jahia.modules</groupId>
-        <version>8.1.5.0</version>
+        <version>8.1.7.0</version>
     </parent>
 
     <artifactId>html-filtering</artifactId>
@@ -84,12 +84,6 @@
             <groupId>javax.jcr</groupId>
             <artifactId>jcr</artifactId>
             <version>2.0</version>
-            <scope>provided</scope>
-        </dependency>
-        <dependency>
-            <groupId>org.apache.felix</groupId>
-            <artifactId>org.apache.felix.framework</artifactId>
-            <version>6.0.5</version>
             <scope>provided</scope>
         </dependency>
         <dependency>
